ARQUITECTURAS E INFRAESTRUCTURAS PARA INTELIGENCIA ARTIFICIAL/Práctica/CodigoPracticaDocker/requirements.txt
psycopg2-binary==2.9.6
requests==2.27.*
psycopg2==2.9.6
uvicorn==0.23.2
pandas==2.0.3
fastapi==0.100.1
typing==3.7.4.3
pydantic==2.1.1
tqdm==4.65.0
xgboost==1.7.6
python-dotenv==1.0.0
SQLAlchemy==2.0.19
